refactor: extract Participation.Sessions from the context facade
First of four modules pulled out of participation.ex, which carried 46 public
functions against 123 private ones -- a ratio no other context in the repo comes
close to. Messaging is nearly as long at 1830 lines, but that length is 72 thin
public functions; Participation was the only context holding an entire private
persistence, orchestration and event-staging layer inside its own facade.

Sessions owns ProgramSession: create, start and edit, the generated-schedule
reconciliation, session persistence, and every query the provider, staff and
admin surfaces render. The functions move verbatim -- this is behaviour
preserving, and the 482 participation tests plus 752 web tests pass with no test
file edited, because the facade's public signatures are unchanged and no test
ever reached a private helper.

complete_session stays behind for now. It writes the session row and sweeps the
roster to absent in one transaction, so it spans two entities and belongs in its
own module once Attendance exists to own the other half -- the rule
provider/offboard_staff_member.ex states for itself.

Three functions are public only to wire modules together, and are marked as
such: persist_lifecycle_update/1 (the session half of completion),
upcoming_scheduled_session_ids/1 (for the roster backfill) and program_name/1
(for the roster read still in the facade, which folds it back to private when it
moves here).

@admin_checked_in_statuses moved to ParticipationRecord.checked_in_statuses/0
first: its three call sites split across Sessions and the future Attendance, and
a fact about record statuses belongs on the record, not duplicated in two
modules.

The move was done by a block-splitter rather than by hand, with two checks that
caught real damage: a round-trip assertion that re-emitting every parsed block
reproduces the file byte-for-byte, and a validator that every @spec names the
function defined beneath it. The second caught nine blocks whose docs had
shifted onto their neighbours -- a multi-line @spec continuation had ended the
scanner early. That shift compiles silently and leaves each function documented
as the one before it.
